Refactor gRPC clients into separate files for readability
